xddd

+++ L02P02_2022-EA-650_2022-RC-652/Controllers/libroController.cs
@@ -136,7 +136,7 @@ namespace L02P02_2022_EA_650_2022_RC_652.Controllers
                             }).ToList();
 
             int cantidadTotal = detalles.Count;
-            double totalVenta = detalles.Sum(d => d.Subtotal) ?? 0;
+            double totalVenta = (double)(detalles.Sum(d => d.Subtotal) ?? 0);
 
             pedido.cantidad_libro = cantidadTotal;
             pedido.total = totalVenta;
